sensor location, external js for maps and chart, rest method for creating sensors

+++ fuel/app/classes/controller/report.php
@@ -30,20 +30,36 @@ class Controller_Report extends Controller
 
     }
 
-    public function action_edit($id) // EDIT NAME/ UNIT
+    public function action_edit($id) // EDIT NAME/ UNIT/ LAT/ LNG
     {
 
         if (Input::post('update'))
         {
+            $val = Validation::forge('validation');
+            $val->add('lat', 'Edit lat')->add_rule('numeric_between',-90,90);
+            $val->add('lng', 'Edit lng')->add_rule('numeric_between',-180,180);
 
-            $sensor = Model_Sensors::find_by_pk(Input::post('sensor_id'));
-            $filters = array('strip_tags', 'htmlentities');
-            $sensor->name = Security::clean(Input::post('name'), $filters);
-            $sensor->unit = Security::clean(Input::post('unit'), $filters);
-            $sensor->save();
+            if ($val->run()){
+                $sensor = Model_Sensors::find_by_pk(Input::post('sensor_id'));
+                $filters = array('strip_tags', 'htmlentities');
+                $sensor->name = Security::clean(Input::post('name'), $filters);
+                $sensor->unit = Security::clean(Input::post('unit'), $filters);
+                $sensor->lat = floatval(Security::clean(Input::post('lat'), $filters));
+                $sensor->lng = floatval(Security::clean(Input::post('lng'), $filters));
+                $sensor->save();
 
-            Session::set_flash('success', 'Name/ unit successfully updated!');
-            Response::redirect('/index.php/report/');
+                Session::set_flash('success', 'Sensor properties successfully updated!');
+                Response::redirect('/index.php/report/');
+
+            }
+            else{
+
+                Session::set_flash('invalid', 'Invalid coordinates! Set latitude between -90 and 90, and
+                                    longitude between -180 and 180!');
+                Response::redirect_back('/index.php/report/', 'refresh');
+
+            }
+
 
         }
 
@@ -72,9 +88,12 @@ class Controller_Report extends Controller
 
             $reports = Model_Reports::find_by('sensorID', Session::get('id'), '=');
 
-            foreach ($reports as $report){
-                $report->delete();
+            if ($reports){
+                foreach ($reports as $report){
+                    $report->delete();
+                }
             }
+
 
              Session::set_flash('success', 'Sensor successfully deleted!');
              Response::redirect('/index.php/report/');
